requirements.txt
openai>=1.0.0
faiss-cpu
rank_bm25
langchain
langchain-openai
langchain-community
fastapi
uvicorn
python-dotenv
pandas
numpy
pydantic
faker
scikit-learn
tiktoken
redis>=5.0.0
httpx
pytest
pytest-asyncio
tenacity
